# packages/server/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CODEXA API Server
  version: 0.2.0
  description: |
    API specifications for the CODEXA Server backend.
    
    This backend provides routes for Clerk authentication lookup, chat session persistence, SSE LLM chat streaming, Polar payment sandbox interfaces, and CodexaLens workspace dependency indexing.
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - BearerAuth: []
paths:
...
  /sessions:
    get:
      summary: List user sessions
      description: Returns a list of all chat sessions associated with the authenticated user.
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SessionSummary'
    post:
      summary: Create a new session
      description: Initializes a new chat session linked to a local project directory workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: User-facing title for the session
                cwd:
                  type: string
                  description: Absolute path of local project directory
              required:
                - title
                - cwd
      responses:
        '201':
          description: Session created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionDetail'
        '400':
          description: Invalid request body
        '402':
          description: Insufficient credit balance
...
  /chat:
    post:
      summary: Start a streaming AI chat completion
      description: Sends message history to the AI model and streams back a SSE response containing text increments, tool execution activities, and final token usage statistics.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: string
                  description: Session ID
                messages:
                  type: array
                  items:
                    type: object
                    description: Input UI messages
                mode:
                  type: string
                  enum: [BUILD, PLAN]
                model:
                  type: string
                  description: Model ID to use (e.g. claude-3-5-sonnet, gpt-4o)
              required:
                - id
                - messages
                - mode
                - model
      responses:
        '200':
          description: Server-Sent Events stream of events
          content:
            text/event-stream:
              schema:
                type: string
        '400':
          description: Invalid request body
        '402':
          description: Insufficient credit balance
        '404':
          description: Session not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SessionSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - createdAt
    SessionDetail:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        cwd:
          type: string
        messages:
          type: array
          items:
            type: object
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - cwd
        - messages
        - createdAt
        - updatedAt
...